MAINT: Split case workflow config to own module
